Fix `top_mem` plugin
Usage of `proc_splitline()` breaks for processes containing spaces as part of the command name, e.g.
```
2566 (tmux: client) S 2459 2566 2459 34817 2566 4194304 290 14 12 0 0 0 0 0 23 3 1 0 124899 11952128 1010 18446744073709551615 94216560173056 94216560738917 140737080092208 0 0 0 0 3674116 134433283 1 0 0 17 14 0 0 0 0 0 94216560947760 94216561019336 94216574795776 140737080101018 140737080101023 140737080101023 140737080102890 0
```
which ends up taking the value of some other field instead of the RSS.

Given that the command name is only used if it is a new topper, avoid using a regex and use
`/proc/[pid]/statm` instead.