docs: check and build dool, not dstat
The PR !30 already fixes the old references in dstat.1.adoc.
But that alone would not be enough to reliably regenerate the doc
file as there is a pre-generated dool.1 shipped.

This fixes the doc Makefile to check and regenerate dool.1 instead
of dstat.1. Without this (and the already open PR !30) checkout and
make of this repo will directly run into:
  make -C docs docs
  make[1]: Entering directory '/tmp/test/dool/docs'
  make[1]: *** No rule to make target 'dstat.1', needed by 'docs'.  Stop.
  make[1]: Leaving directory '/tmp/test/dool/docs'

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>