Fix and make project structure more Python packaging idiomatic
The project structure and the pyproject.toml would only cause the
dool.py script to be installed, not the plugins.

To make it easier for downstream packages, like Gentoo in my case,
dool should use an idiomatic Python project structure. This change
moves the files accordingly, adds a __init__.py (as recommended by
flit), and adjusts pyproject.toml accordingly.