Fix empty row for the first iteration of --top-io
Stats from non-root processes cannot be read and should be skipped. On
the first iteration, we historically reported the processes with the
most I/O, so let's restore that functionality.

Reduce memory consumption by not storing unused fields, and skip reading
process names when the counters cannot be read anyway. Refactor the code
to make it easier to review, some exceptions could never occur.

Fixes: db330372733b ("Skip files we cannot get /proc/$PID/io data from (Fixes #94?)")